fix(rag): address Phase B code review findings
- Add input validation on public method boundaries (IterativeRetriever,
  SufficiencyEvaluator, AnswerFaithfulnessEvaluator)
- Split RagOrchestrator into partial class (RagOrchestrator.MultiHop.cs)
  to stay within 400-line file limit
- Wire FaithfulnessConfig.HallucinationThreshold into faithfulness check
  instead of relying solely on LLM's is_faithful verdict
- Fix wasted double-retrieval in ExecuteRoutedPipelineAsync by moving
  CRAG check before retrieval+rerank steps
- Fix inaccurate XML doc on IterativeRetriever re-retrieval behavior
- Replace null-forgiving operator with explicit guard in multi-hop path
